BUG#34703698 after_commit hook may be invoked out of order even when binlog_order_commit=1
Background
-------
When binlog_order_commit=1, for any two transactions and for any sub-step of
the commit phase, it is expected that the transaction that was written to the
binary log first, executes the sub-step first.

This is ensured in two ways, depending on whether the two transactions belong
to the same BGC group or not:
- When two transactions belong to the same BGC group: the leader of the BGC
group traverses the group's queue in the same order during all sub-steps.
- When two transactions belong to different BGC groups: LOCK_commit is held
for the duration of the commit phase, so only one BGC group at a time can
execute. BGC groups enter the commit phase in correct order due to the logic
in MYSQL_BIN_LOG::change_stage and functions called from it.

Problem
-------
The leader thread releases LOCK_commit after executing most of the
sub-steps of the commit phase, but before executing the after_commit hook. It
is possible that the leaders of BGC groups N and N+1 call the after_commit
hooks for their respective groups in parallel, or even that the leader for
BGC group N is scheduled out for a long time and the leader of BGC group N+1
calls the after_commit hook first.

This is supposedly not a problem for the currently existing components that
register callbacks for the after_commit hook. However, it seems to violate
the expectations of setting binlog_order_commits=1.

Analysis / Root-cause analysis
------------------------------
Leader threads of BGC groups leaving commit queue can be interleaved
and execute after commit hook in changed other order.

Solution
--------
Introduce a new BGC stage and use change_stage instead of
mysql_mutex_unlock(&LOCK_commit) between the calls to
process_commit_stage_queue and process_after_commit_stage_queue in
MYSQL_BIN_LOG::ordered_commits. Change MYSQL_BIN_LOG::ordered_commits
and add AFTER_COMMIT_STAGE.

Change-Id: I35765053d92fbb8ec912ffb75240c7ac02cf665c