BUG#34768191 - Implement 'firstMatch' optimization in SPJ
Patch introduce 'firstMatch' optimizations into the SPJ block.

Depending on any parallelism statistics being available for a
SCAN_FRAGREQ, the SPJ block will only send such REQs to a subset
of the table fragments in parallel.

Iff the table being scanned is part of a semi- or anti-join,
the EXISTS / NOT EXISTS outcome of th equery is known as soon
as the first 'match' on a specific key/range is found. Thus we
may skip that key from the remaining SCAN_FRAGREQ's not yet sent.

This patch takes advantage of that by implementing the method
Dbspj::removeMatchedKeys() where we rebuild the range keys to
only contain keys not having a match yet. Note that in case the
SPJ query also use a parameterized pushed condition on this
specific table, the parameter value related to the key need
to be removed as well.

As we may possible need to to resent the full set of range scans
on a table (Part of batch overflow handling) we need to save away
the original keys and parameters before anything is 'removed'.
m_rangeCntSave, m_rangePtrISave, and m_paramPtrISave is introduced
for that.

Dbspj::scanFrag_parent_batch_repeat() use these to reset the
REQuest state to its original contents if needed.

The TPC-H queries Q4, Q21 and Q22 takes great advantage of
this optimization. Some measurements from running these
test on a 4 datanode x 2 LDM's config:

Elapsed query time:

  - Q4:  0,38 -> 0.19s
  - Q21: 1,91 -> 1.00s
  - Q22: 0,52 -> 0.27s

Inspecting the ndbinfo.counters 'where block_name='DBSPJ'' reveals a
similar effect on the amount of data being processed by SPJ and the LDMs.
The counter SCAN_ROWS_RETURNED is reported below, but similar effect is
seen on a couple of the other counters as well:

  - Q4:  181303 -> 133274
  - Q21: 782869 -> 472717
  - Q22: 362922 -> 148813

Change-Id: I4df5bbf297ea78a08257c3d8dfa6a89ac1eecdbc