Bug #34699398 Row estimates for joins ignores histograms
There is a function EstimateFieldSelectivity() that estimates the
selectivity of a given field (when there is an equi-join with some
other field.) As it was, this function looked for (non-hash) indexes
starting with that field, and used KEY::records_per_key() to estimate
the selectivity. If no suitable index was found, a default selectivity
of 10% was used.

This patch ensures that if there is no suitable index, we will look
for a histogram. If there is one, we use
Histogram::get_num_distinct_values() to estimate the selectivity.

Change-Id: I83af46e04142aa0260e701cd07e03c246696b3be