Bug#33100586 CRASH IN ITEM_SINGLEROW_SUBSELECT::STORE
Issue:
If the subquery of an order by clause has a view reference, it is
not cleaned up. This is because the subquery might be referred to
again during the execution.

In the case of execution of such a prepared statement, the server
exits as the query expression from the order by clause is present
but not marked optimized.

Fix:
Cleanup query expression of subquery if it is marked as not used
by the resolver.

Change-Id: I7ce2c6243e132339ad69c2c353f64c8d1c0485a2