Bug #32762229 [ERROR] [MY-013183] [INNODB] ASSERTION FAILURE: ROW0MYSQL.CC:2928:PREBUILT->TEMP
A DELETE statement which is executed using an index skip scan fails
an assertion on the row template type. The index skip scan
initialisation sets a keyread flag to indicate that it is sufficient
to read only the indexed fields. Based on the keyread flag, a
template is built which specifies a subset of the fields instead of
the whole row. While performing the row delete, this hits an assert
since the template type is required to specify the whole row for
deletes.

This is fixed in index skip scan initialisation by checking if the
query requires fields which are not included in the index and
setting the keyread flag accordingly. A test case is added for a
skip-scan delete query.

Change-Id: Iaa0cbd380ae12519b3dadc7924b538432bf6a77d