Bug#34768106 SPJ: Collect better scan 'parallelism' statistic
Addendum patch#2

Introduce scanFrag_getBatchSizes() to estimate
how many more rows we may fetch into the available client batch
buffers, given the configured 'bytes' and 'rows' available. We are also
limited by the 12-bit correlation id range. Note that some 'bytes', 'rows'
and 'correlationIds' may already have been used up as reflected in 'data'.

The collected 'record size'-statistic is used to estimate how many 'rows'
we may fit in the available batch bytes. The most restrictive limitation
of 'bytes', 'rows' and 'correlationId' is used to calculate the 'BatchSize'
we may use which is returned from this method.

This allowed the calculation of 'available*' batch buffers space
to be collected into a single method.

scanFrag_parallelism() is changed to take the BatchSize calculated
by scanFrag_getBatchSizes() as argument, instead of the
availableBatchBytes and availableBatchRows.

Reset of data.* counters (where relevant) is moved to before
calculation of 'parallelism' in order to make the fill batch size
available in calculation (No m_totalRows and m_totalBytes from
previous rounds remaining as 'used')

Change-Id: I574e9b197d9d134d8c4cb524334f356e4d235959