Bug#34806320 connection sharing fails if client via unix-socket
When connection sharing is enabled and a connection is made via
unix-sockets, the server-connection isn't shared.

For connection-sharing to work, the router must be able to fetch the
clients password, but for unix-sockets the password is never requested.

Currently, the password is requested after the TLS handshake, but in the
case of unix-sockets, the transport is already secure without TLS.

Change
------

- request client's password even with TLS if connection is over
  unix-sockets

Change-Id: I85653cc620681f5384b3f179b0b2051523985a0f