Bug#34796631: Hypergraph doesn't consider indexed_col = function as sargable [1/2, functions]
The hypergraph optimizer currently creates index lookups only for
predicates on the form column = constant (like t1.x = 3), or simple
join conditions on the form column = column (like t1.x = t2.x).

This patch makes it produce index lookups for more kinds of
predicates:

1) For table predicates column = constant, it now also accepts
expressions with subqueries that are constant during the execution.
For example, WHERE t1.x = (SELECT MAX(x) FROM t2) could now become an
index lookup if there is an index on t1.x. Also, constant expressions
that invoke stored functions are considered sargable and can result in
index lookups. (Because we don't want the hypergraph optimizer to
execute subqueries or stored procedures, even if they are constant,
ref_lookup_subsumes_comparison() is changed so that it does not
evaluate these constant expressions. This may cause a redundant filter
to be placed on top of the index lookup.)

2) For join predicates, it now accepts equalities where one side is a
reference to an indexed column and the other side is a function of
columns in one other table. For example, a join condition such as
WHERE t1.x = t2.x + 1 may become a nested loop join with t2 as the
outer table and an index lookup on t1 as the inner table, given that
there is an index on t1.x.

Change-Id: Iac76f7c2c523aa493813d57d09fb64d0d3059ca4