Bug#34782389: regression: match against: assert_consistent_hidden_flags: Assertion `item->hidden == hidden' failed.
An assert failure was seen in a query that had an IN subquery which
performed full-text search.

The assertion was triggered when adding a hidden element in the SELECT
list for full-text functions in the HAVING clause. The assertion
detected that the item was already in the SELECT list. (The query had
no explicit HAVING clause, but the in-to-exists transformation had
added one implicitly.)

Fixed by adding the full-text function to the SELECT list only if it
is not already there.

Change-Id: I71c1cb3cfabe4df6ac76d13a432386dc0abb45f1