Bug#34768353 Reduce number of REQuest roundtrips to fetch sorted-SPJ-result
Preparation patch#2, refactor:

Main patch will allow an ORDERED SPJ result to self-request more
rows from an ORDERED rootTreeNode, which is operating with a
row_batch_size=1.

In such requests the child-treeNodes will already have received some
result rows -> data.m_totalRows, *.m_totalBytes and *.m_corrIdStart
will reflect the state of the rows already received.

We need these values to be kept, and taken into consideration
when calculating available batch space, and 'parallelism' to be
used for any more requests to be made. Thus, we can not reset these
counters as part of each (child-)scanFrag_send() anymore.

This patch moves the reset of these counters reflecting the clients
side batch states to the batch-cleanup logic instead. An extra 'done'
argument is introduced which is set to 'true' if the completed 'batch'
was the last one in this round. (Still always 'true' until main part
of this patch set is introduced)

Change-Id: Ia6cdae17eaeefb8aad1f2a665280176c7a465a7b