Bug#34768353 Reduce number of REQuest roundtrips to fetch sorted-SPJ-result
Introduce self-requests of a NEXTREQ for T_SORTED_ORDER result set.

When a batch of results has completed, Dbspj::batchComplete() detects
that we have completed a (single row) fetch from a T_SORTED_ORDER
result set. If the ORDERED rootTreeNode is still the only 'active'
treeNode in the request, and we have remaining space in the batch buffer,
we may immiately issue a NEXTREQ retrieving more result rows.

Note that this patch uncovers a bug in the NdbQueryOperation result handling.
The TPC-H Q13 result set will differs with ndb_join_pushdown=on/off.
(Fixed in a later patch).

Change-Id: I77d45f73ac37b895d902fb43a92f2d39c5d400c7