Bug#34717171: Hypergraph :Assertion `false' failed               in join_optimizer.cc
When creating equalities from multiple equalities,
hypergraph finds replacements from tables which are
not visible for a join i.e it finds replacements from
the inner side of semijoin which results in seeing
the semijoin conditions that break the rules. Such
semijoin conditions result in plans which have different
costing leading to an assert failure.
Solution is to find the correct replacements when
concretizing multiple equalities.

Change-Id: I0434f4bc15ab35bf670bd59e673f962199deb1a7