Bug#34823366 possible data-race creating x-connections
TSAN reports a possible data-race between creating a x-connection and
further handing the connection.

The connection is created in the acceptor's io-thread until it blocks,
but then gets futher handled in the io-ctx of the client-socket ...
which may be another io-thread.

It also blocks the acceptor until the x-connection would-blocks (which may
take a long time if all data can be read from the socket directly),
resulting in a slower accept-rate.

Change
------

- defer handling the x-connection

Change-Id: I1323283ad0f05cbc82ec85274545099705749d83