Bug#34288890: Histogram commands invalidate TABLE_SHARE
The ANALYZE TABLE UPDATE/DROP HISTOGRAM commands invalidate the
TABLE_SHARE. After the TABLE_SHARE has been invalidated subsequent
queries must wait for all currently running queries to terminate before
the old TABLE_SHARE can be freed and a new TABLE_SHARE can be
initialized with the updated collection of histograms for the table.
Only when these operations are completed can new queries referencing the
table begin acquiring TABLE objects and proceed executing. This has the
potential to introduce long waits, as queries issued after the
TABLE_SHARE was invalidated must wait for any existing long-running
queries that reference the old TABLE_SHARE to terminate.

This patch changes the behavior of the histogram commands to mark tables
for reopening instead of invalidating the TABLE_SHARE. Instead of having
a single set of table histograms cached on the TABLE_SHARE we maintain a
collection of reference-counted sets of table histograms on the share.
When histogram commands modify the histograms on a table we insert a new
snapshot of the set of histograms into the collection on the TABLE_SHARE
and mark it current. When a TABLE object is opened it acquires a pointer
to the current snapshot of the set of histograms for the table from
the share, and when the TABLE object is closed it releases its pointer
back to the share.

By having multiple reference-counted versions of histogram statistics
for a table we avoid the potential wait of synchronizing all queries on
the table around the reinitialization of the TABLE_SHARE when histograms
are updated or dropped.

Change-Id: Ida527179b6aa8d74ec550537b09891ddf3777cb2