Bug#34821222: Hypergraph: mysqld crash-signal 11 - IsAnd & CommonSubexpressionElimination
Common subexpression elimination fails when replanning a query without
the IN-to-EXISTS transformation. The reason is that always false
conditions were removed from an Item_cond_or by remove_eq_conds() in
the first round of planning (with IN-to-EXISTS transformation),
leaving only an empty OR condition. CommonSubexpressionElimination()
expects Item_cond_or to have at least one argument, accesses the first
argument unconditionally, and fails when there is no argument.

It is very similar to Bug#34503695, which was fixed by making
remove_in2exists_conds() copy the condition structure so that the
modifications done in the first planning were not seen by the second
planning. That fixed the problem with always false conditions in the
WHERE clause and HAVING clause. In this new bug, however, the always
false conditions are in a semijoin condition, which is not touched by
remove_in2exists_conds().

To fix the new bug, CommonSubexpressionElimination() is taught to
handle empty Item_cond_or objects as constants with the value FALSE.

This fix makes the original fix for Bug#34503695 redundant, so
remove_in2exists_conds() is changed back so that it no longer copies
the condition structure. It is also simplified by using some existing
helper functions.

Change-Id: I7b674007a754dc9f289118a31a68c13310a67b89