Bug#34449016 since 8.0.23+ CREATE USER became slower
Problem
=======
Creation of 20k users with the IP in the hostname takes over 10 hours comparing
to 12 minutes before the WL#14074.

Analysis
========
WL#14074 changed the way ip_mask value of the ACL_HOST_AND_IP sturcture is set, when
the IP is used in the hostname part. Previously it was set, only when the mask value
was provided. Because of this bug, IP provided with the host name was treated
as the wildcarded hostname. The algorithm for building a structure that keeps all
wildcards is performance costly.

Fix
===
A host part of the user account is treated as a wildcard, only if the mask is not
255.255.255.255.

Change-Id: I6e07f0cc73911951485024bda152920e87c485f9