Bug#34796631: Hypergraph doesn't consider indexed_col = function as sargable [2/2, hyperpredicates]
The hypergraph optimizer currently does not create index lookups for
equality predicates where one of the sides references more than one
table. For example, t1.x = t2.x + t3.x does not result in an index
lookup on t1, even if there is an index on t1.x.

Hash joins also handle this kind of predicate suboptimally. Hash joins
between t1 and {t2,t3} will not make a proper hash join predicate of
it, but will instead make it a so-called "extra" predicate evaluated
after the join, which means the hash join essentially degenerates to a
nested loop join.

This patch fixes both of these issues. If there is an index on t1.x,
the hypergraph optimizer may now propose a nested loop join with t2
and t3 on the outer side and an index lookup on t1 on the inner side,
when it sees the above mentioned join predicate. And, regardless of
whether there is an index, it may propose a proper hash join between
t1 and {t2,t3} where a hash table is built either for t1.x or for the
expression t2.x + t3.x.

Change-Id: Ie88602baaf03db6e6356847b0cbf0e9796217a11