Bug#34828328: Assertion `right != nullptr' failed in               make_join_hypergraph.cc
This is a regression from Bug#34717171. The patch restricted finding
candidates from only visible tables for a join when creating equalities
from a multiple equality. For the case in bugpage, we have a condition
where we cannot find replacements in the visible tables of the join,
but outside of it. So another case of rule breaking semijoin conditions
not able to find replacements.
Since it is only right to find the replacements from visible tables of
a join and also that rapid does not support seeing inner tables of a
semijoin from outside the semijoin, we continue to find the candidate
in visible tables of a join. However, when we fail to find the candidate,
then we try to find in all tables in a join.

Change-Id: Ibca8bc0bb0996674f230792e9f222964b080d90a