Bug#33725542: A  bug in sql_derived.cc cause mysql-server crash               when input some sql statement..
When a call to create_materialized_table() is made, it checks
if the table is already created and if so it presumes that
if the derived table is marked constant, this table should have
a NULL row. However, for a CTE which uses shared materialization
it might not be true.
For ex:
If a CTE which does not use any tables is part of the FROM
clause and is in the SELECT list, optimizer materializes the
derived table when FROM clause is processed. When it tries
to create materialized table for the SELECT LIST, it checks
if one of the CTE references has materialized the table. If
so, it opens the temp table. However, there is a where
condition present in the subquery of the SELECT LIST which
makes it a const table. So the assert which expects that a
constant derived table if already created should have a null
row is not met and fails.
Solution is to not do the checks w.r.t const derived tables
when shared materialization is used for a CTE, since it no more
holds good.

Change-Id: I2f39db0b4d76ae5b65df72966cfda34cd6c02404