Bug #34467659: "SHOW FIELDS FROM" fails against a view when a view accessing from a view is recreated, and the user grants a table level privs via role
Description:
============
View v1 selects from table t1 and view v2 selects from v1.
When v1 is dropped and then created again in the same way, SHOW FIELDS
FROM v1
reports an invalid view error. This issue is similar to Bug #34341533,
but the difference is that in this case the grant is on *.* instead of
on specific resources. It doesn't matter if the grant is done via role or directly to a user.

Analysis:
=========
On a view recreation, all views selecting from it are updated and set to
valid if applicable. In more detail, on v1 recreation
open_views_and_update_metadata is called which in turn calls
check_view_privileges for all views using v1 (v2 in this case)
and it loop over all tables used by v2 (t1 in this case),
calling check_single_table_access for them. check_single_table_access
doesn't deny access, but also doesn't set grant privileges on table
level because of the below condition:

if (!(all_tables->belong_to_view &&
     (thd->lex->sql_command == SQLCOM_SHOW_FIELDS)) &&
      check_grant(thd, privilege, all_tables, false, 1,
      no_errors))
   goto deny;

 (the first part of the condition is false, so check_grant is not
 called)

 Then Query_expression::check_materialized_derived_query_blocks "walks"
 over t1  columns checking if there are column level privileges
 (Item::check_column_privileges) and it fails because there are no
 column level privileges. Note, that check_column_privileges will succeed
 if there are table level privileges, but they are not set as the above
 check_grant was not called. As the result an error occurs and "valid"
 property is not set to v2 and further SHOW FIELDS will fail.

 Fix:
 ====
 Remove the described condition above and call check_grant in
 check_single_table_access unconditionally.
 Note, that according to the manual:
 "SHOW COLUMNS displays information only for those columns for which you
 have some privilege.", so the check on table level or column level must
 succeed. The described condition causes omitting table level check.
 The check is hit when creating v1 for the second time (in fact
 thd->lex->sql_command == SQLCOM_SHOW_FIELDS is true then). It is not
 hit when SHOW FIELDS is called. And
 thd->lex->sql_command == SQLCOM_SHOW_FIELDS
 is false when processing any other commands.

Change-Id: Ie610d2784d4bbe516fe10612c65a829a7f5651c7